Changing methods in the PreferencesHandler and separating tests into different files
