+++ Woof/WoofTests/Model/Preferences/PreferencesHandlerTestHelper.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct PreferencesHandlerTestHelper {
+    static func cleanStorage() {
+        let keyValueStorage = KeyValueStorage(KeyValueStorage.Name.preferences)
+        keyValueStorage.deleteData(for: KeyValueStorage.Key.userPreferences)
+    }
+}
